source/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
aiofiles
python-dotenv